+++ cli/main.py
@@ -7,5 +7,4 @@ app.command(name="val")(val.run)
 
 if __name__ == "__main__":
     app()
-
-
+
